+++ changelog.md
@@ -1,6 +1,5 @@
 # 1.0.6
-- Add android support
-(Note that this is the first version released on the Geode index.)
+- Idk yet
 
 # 1.0.5
 - Make it so the blinking eye slowly moves into the screen before blinking, so you know it's coming.